requirements.txt
discord.py>=1.7.1
pyyaml>=5.4.1
pymongo>=3.11.3
